Consolidate adjacent free cells into a larger free cell
This changes the layout of the cell header, and adds two different overlays for
allocated and freed cells. We now have two lists instead of one:

1. a doubly-linked list of all adjacent cells, and
2. a singly-linked list of all free cells.

The second list steals a word from the cell's data payload, which it can safely
take because it only deals with unallocated, free cells.

Fixes #1
Fixes #2